Refactor speech recognition
